# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RemessasSteam API",
        "description": "API para gestão de projetos, parceiros e divisão de receitas",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reports/{report}/allocate": {
            "post": {
                "tags": [
                    "Reports"
                ],
                "summary": "Alocar projetos no relatório",
                "operationId": "a21689c7021442c5b1b88a17e6679a22",
                "parameters": [
                    {
                        "name": "report",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "allocations"
                                ],
                                "properties": {
                                    "allocations": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "project_id": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "project_net_amount": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 500
                                                },
                                                "currency": {
                                                    "type": "string",
                                                    "example": "USD"
                                                },
                                                "units_sold": {
                                                    "type": "integer",
                                                    "example": 100
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "overwrite": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Alocações salvas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}